requirements/dev.txt
-r base.txt

black==24.4.2
django-debug-toolbar==4.3.0
flake8==6.0.0
isort==5.12.0
